move `htmlPlugin` to `render-html-plugin.js`
